+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Post-release panic
 **Added**: 
 - JER integration for liquid death crystal drops
 - Hot Reloading of configs
+- Comments to overlay multiplier config options
+- Purity background to AspectFromItemStackCategory (lighter the more % of the item's aspect this one is)
+
+**Fixed** Off-by-one AspectFromItemStackCategory background
 
 ## 1.0.0:
 First release